[DITP-pilotage/pilote-2] refactor: mutualise SegmentedControl dans kpilote-ui
Factorise le dernier composant restant. Contrôle canonique = version webapp
(ToggleGroup autonome, générique, icônes, non-déselectionnable) déplacée dans
@pilote/kpilote-ui/SegmentedControl, avec support aria-label OU aria-labelledby.

L'usage « champ de formulaire » de l'admin (label + Field, ex-RadioGroup) est
reconstruit via un wrapper mince SegmentedField (Field + contrôle partagé,
associé par aria-labelledby), qui conserve l'API { label, value, onValueChange,
options } des call-sites admin.

Le comportement déselectionnable de l'ancien RadioGroup admin n'était utilisé
nulle part, il n'est donc pas repris.

Refs:
Project: pilote-2
Client: DITP-pilotage
Category: refactor
Stack: react, typescript
Tags: design-system, monorepo
Files-changed: 7
